task yalv configure react navigation
