+++ client/components/authenticate/index.js
@@ -1,0 +1,1 @@
+export { LoginScreen } from './login';
