+++ package.json
@@ -32,6 +32,7 @@
     "nodemon": "^1.19.0",
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
+    "react-navigation": "^3.11.0",
     "react-redux": "^7.0.3",
     "redux": "^4.0.1",
     "redux-actions": "^2.6.5",
